\name{nmbu-package}
\alias{nmbu-package}
\docType{package}
\title{
Simplified loading and updating of RcmdrPlugin.NMBU
}
\description{
Simplified loading and updating of RcmdrPlugin.NMBU
}
\details{
Simplified loading and updating of RcmdrPlugin.NMBU
}
\author{
Kristian Hovde Liland
}
\keyword{ package }
